+++ tinybaker/__init__.py
@@ -2,5 +2,5 @@ from .transform import Transform
 from .combinators import sequence, merge, map_tags
 from .workarounds import fileset
 from .context import BakerContext
-from .namespace_transforms import InputTag, OutputTag
+from .tag import InputTag, OutputTag
 from .cli import cli
