+++ tinybaker/namespace_transforms.py
@@ -2,9 +2,7 @@ from typing import Dict, Any
 import contextvars
 from .fileref import FileRef
 from .exceptions import BakerError
-
-input_files_ctx = contextvars.ContextVar("input_files")
-output_files_ctx = contextvars.ContextVar("output_files")
+from .tag import InputTag, OutputTag
 
 
 def namespace_to_transform(BaseClass, source_ns):
@@ -44,34 +42,3 @@ def dict_to_transform(BaseClass, ns_dict: Dict[str, Any], name_outer=None):
             cls._ns_dict["script"]()
 
     return DerivedTransform
-
-
-class BaseTag:
-    FILE = "file"
-    FILESET = "fileset"
-
-    def __init__(self, name, in_or_out):
-        self.name = name
-        self.in_or_out = in_or_out
-
-    @property
-    def target(self):
-        raise BakerError("Cannot reference target outside of a TinyBaker context!")
-
-
-class InputTag(BaseTag):
-    def __init__(self, name):
-        super().__init__(name, "in")
-
-    @property
-    def ref(self):
-        return input_files_ctx.get()[self.name]
-
-
-class OutputTag(BaseTag):
-    def __init__(self, name):
-        super().__init__(name, "out")
-
-    @property
-    def ref(self):
-        return output_files_ctx.get()[self.name]
